Add : DIe 애니메이션 추가

+++ Assets/Resources/ScritpableObjects/Player/PlayerSO.cs
@@ -10,7 +10,8 @@ using Utill;
 [CreateAssetMenu(fileName = "Player", menuName = "Characters/Player")]
 public class PlayerSO : ScriptableObject
 {
-    public float Health;
+    public int Health;
+
     [field: SerializeField] public TargetSearchData TargetSearchData { get; private set; }
     [field: SerializeField] public PlayerGroundData GroundData { get; private set; }
     [field: SerializeField] public PlayerAttackData AttackData { get; private set; }
